FE-Feat: 예약 시트 컴포넌트 분리 및 반응형으로 사이드바, bottomsheet 나타나도록 구현
